requirements.txt
streamlit==1.29.0
pycryptodome==3.19.0
Pillow>=10.4.0
numpy==1.26.2
opencv-python
